src/api/requirements.txt
fastapi==0.68.0
psycopg2-binary==2.9.1
gunicorn==19.9.0
pandas==1.3.3
alembic==1.4.3
pytest==6.2.5
requests==2.26.0
uvicorn==0.14.0
python-multipart==0.0.5
fastapi-cache2==0.1.8
itsdangerous==2.0.1
tortoise-orm==0.19.0
asyncpg==0.25.0
